Bugfixes and observation fixes

+++ src/app/redux/auth/logout/actions.js
@@ -38,10 +38,15 @@ export const logoutSuccess = (userData) => {
 }
 
 export function onLogout(){
-    return (dispatch, getState) => {
-        userManager.removeUser();
-        dispatch(remove(USER_LOCALSTORAGE, onClear));
-        userManager.signoutRedirect();
+    return (dispatch) => {
+        let localStorageData = JSON.parse(localStorage.getItem(USER_LOCALSTORAGE));
+        if (localStorageData && localStorageData.data && localStorageData.data.access_token) {
+            userManager.removeUser();
+            dispatch(remove(USER_LOCALSTORAGE, onClear));
+            userManager.signoutRedirect();
+        } else {
+            dispatch(push(Path.root))
+        }
     }
 }
